+++ requirements.txt
@@ -2,6 +2,7 @@ samplotlib==0.0.6
 emcee==3.1.4
 transformers==4.36.0
 openai==1.12.0
+anthropic==0.16.0
 fschat==0.2.23
 peft @ git+https://github.com/huggingface/peft.git@b1059b73aab9043b118ff19b0cf96263ea86248a
 fire==0.5.0
@@ -11,6 +12,5 @@ matplotlib
 numpy
 pandas
 getdist
-anthropic
 ninja
 rwkv
